requirements.txt
boto3==1.19.11
botocore==1.22.11
moto==2.2.13
pip==21.1.1
prettytable==2.4.0
pytest==6.2.5
python-dotenv==0.19.1
